+++ middlewares/activity.validations.js
@@ -1,20 +1,21 @@
 const { body } = require("express-validator");
 const db = require("../models");
 
-module.exports = [
+const validationsActivities = [
   body("name")
-    .notEmpty()  // Valid that it is not empty
+    .notEmpty() // Valid that it is not empty
     .withMessage("Input required")
     .bail()
     .not()
-    .isNumeric()  // Valid that it is a string (I place the .not())
+    .isNumeric() // Valid that it is a string (I place the .not())
     .withMessage("The value cannot be numeric.")
-    .isLength({ min: 5 })  // Valido su longitud
+    .isLength({ min: 5 }) // Valido su longitud
     .withMessage("String must be more than 4 characters")
-    .bail() 
+    .bail()
     .isAlphanumeric() // Valid that it is Alphanumeric
     .withMessage("Special characters are not accepted")
-    .custom(async (value) => {  // Function that detects if an activity with the same name exists
+    .custom(async (value) => {
+      // Function that detects if an activity with the same name exists
       const activity = await db.Activity.findOne({ where: { name: value } });
       return (
         activity &&
@@ -22,11 +23,13 @@ module.exports = [
       );
     }),
 
-  body("content") 
+  body("content")
     .notEmpty() // Valid that it is not empty
     .withMessage("Input required")
     .bail()
     .not()
-    .isNumeric()  // Valid that it is a string (I place the .not())
+    .isNumeric() // Valid that it is a string (I place the .not())
     .withMessage("The value cannot be numeric."),
 ];
+
+module.exports = { validationsActivities };
