+++ routes/activities.js
@@ -1,12 +1,13 @@
 const router = require("express").Router();
 
 // Middleware validations required
-const activityValidation = require("../middlewares/activity.validations");
+const {validationsActivities,indexValidation} = require("../middlewares/index.validations");
+
 
 // Controller Api required
 const { store } = require("../controllers/activity.controller");
 
 // Routes
-router.post("/", activityValidation, store);
+router.post("/", validationsActivities, indexValidation, store);
 
 module.exports = router;
